Return Promise from `editRpc`
This action creator returned a thunk that performed async work, but it
didn't return a Promise. Now it does.

In the one place where this is used, it didn't seem important to update
the callsite to block on this finishing. Only one call followed it in
the event handler, and it didn't seem to depend on this.