Return Promise from `setFirstTimeFlowType` thunk
`setFirstTimeFlowType` returned a thunk that didn't return a Promise,
despite doing async work. It now returns a Promise.

The callers have been updated to `await` the completion of these
operations because a `history.push` immediately followed. That may have
resulting in the operation being interrupted if it wasn't `await`-ed.
I'm also not sure it would have been appropriate to proceed with the
redirect if the operation failed.